Packet 169 - System Broadcast (#47)
This packet is sent from the server to send a broadcast to all logged in clients.

It features two attributes: An unknown int and a message.

As noted in the comments, I am not sure what the int sent does. I looked at decompiled code and it seems to make it skip over reading a string if set to zero or something. However even when set to zero it still displays.